+++ Flurry Analytics API Documentation/dir_37af60843fee685f06bcbfe2d1667c15.js
@@ -1,4 +1,0 @@
-var dir_37af60843fee685f06bcbfe2d1667c15 =
-[
-    [ "PublicAPIs", "dir_b196872496082968a63013e0c67815ed.html", "dir_b196872496082968a63013e0c67815ed" ]
-];